% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxcdf.R
\name{pbvn}
\alias{pbvn}
\title{Cumulative Distribution Function for Bivariate Normal Distribution}
\usage{
pbvn(q1, q2, rho)
}
\arguments{
\item{q1}{Threshold or upper bound for the first variable.}

\item{q2}{Threshold or upper bound for the second variable.}

\item{rho}{Correlation between the two variables (must be between -1 and +1).}
}
\value{
The CDF (probability that both variables drawn from a standardized bivariate
normal distribution parameterized by `rho` will be lower than the corresponding values
in `q1` and `q2`).
}
\description{
Calculates the CDF (cumulative distribution function) of a
2D/bivariate standardized normal distribution using a fast approximation
based on the 'erf' function.

This is faster than the more exact method from Drezner used for 2D problems
in \link{pmvn}, but it is much less precise (error can be as high as \eqn{10^{-3}}).
}
\details{
Note that this function will not perform any input validation. It is up to
the user to check that the data passed here is valid.
}
\examples{
library(approxcdf)

### Short example problem
q1 <- 0.25
q2 <- -0.25
rho <- 0.5

### Calculate an approximate CDF for this distribution
pbvn(q1, q2, rho)

### Compare against Genz's more precise method
### (exact to more than 8 decimal places)
if (require(mvtnorm)) {
    pmvnorm(upper=c(q1,q2), corr=matrix(c(1,rho,rho,1), nrow=2), algorithm=TVPACK())
}
}
\references{
Tsay, Wen-Jen, and Peng-Hsuan Ke.
"A simple approximation for the bivariate normal integral."
Communications in Statistics-Simulation and Computation (2021): 1-14.
}
